When marking a comment as ham (not spam) we must update the state
We also mark a comment as not spam when it's been approved. After some thoughts we consider that we must not mark the comment as spam nor ham when it's rejected, as we can't be sure if it's rejected because it's spam or because the content is not appropiated for us (i.e. an undesired controversial comment)